Se agrgan cambios para el cálculo de vacaciones Por régimenLaboral, se hacen correcciones y se edita el contexto para agregar campos en las vistas

+++ swTH/bd.swth.entidades/Negocio/SolicitudVacaciones.cs
@@ -22,7 +22,7 @@ namespace bd.swth.entidades.Negocio
         public int? IdEmpleadoReemplazo { get; set; }
 
         [NotMapped]
-        public int DiasVacaciones { get; set; }
+        public decimal DiasVacaciones { get; set; }
         
         public virtual Empleado Empleado { get; set; }
