+++ swTH/bd.swth.entidades/ViewModels/SolicitudVacacionesViewModel.cs
@@ -28,7 +28,7 @@ namespace bd.swth.entidades.ViewModels
         public string RazonNoPlanificado { get; set; }
         public bool RequiereReemplazo { get; set; }
 
-        public int VacacionesAcumuladas { get; set; }
+        public decimal VacacionesAcumuladas { get; set; }
 
         public List<SolicitudPlanificacionVacaciones> ListaPLanificacionVacaciones { get; set; }
